// portal/proto/isuxportal/resources/notification_pb.d.ts
export declare type Notification = Message<"isuxportal.proto.resources.Notification"> & {
  /**
   * @generated from field: int64 id = 1;
   */
  id: bigint;

  /**
   * @generated from field: google.protobuf.Timestamp created_at = 2;
   */
  createdAt?: Timestamp;

  /**
   * @generated from oneof isuxportal.proto.resources.Notification.content
   */
  content: {
    /**
     * @generated from field: isuxportal.proto.resources.Notification.BenchmarkJobMessage content_benchmark_job = 3;
     */
    value: Notification_BenchmarkJobMessage;
    case: "contentBenchmarkJob";
  } | {
    /**
     * @generated from field: isuxportal.proto.resources.Notification.ClarificationMessage content_clarification = 4;
     */
    value: Notification_ClarificationMessage;
    case: "contentClarification";
  } | {
    /**
     * @generated from field: isuxportal.proto.resources.Notification.TestMessage content_test = 5;
     */
    value: Notification_TestMessage;
    case: "contentTest";
  } | { case: undefined; value?: undefined };
};

/**
 * Describes the message isuxportal.proto.resources.Notification.
 * Use `create(NotificationSchema)` to create a new message.
 */
export declare const NotificationSchema: GenMessage<Notification>;

/**
 * @generated from message isuxportal.proto.resources.Notification.BenchmarkJobMessage
 */
export declare type Notification_BenchmarkJobMessage = Message<"isuxportal.proto.resources.Notification.BenchmarkJobMessage"> & {
  /**
   * @generated from field: int64 benchmark_job_id = 1;
   */
  benchmarkJobId: bigint;
};

/**
 * Describes the message isuxportal.proto.resources.Notification.BenchmarkJobMessage.
 * Use `create(Notification_BenchmarkJobMessageSchema)` to create a new message.
 */
export declare const Notification_BenchmarkJobMessageSchema: GenMessage<Notification_BenchmarkJobMessage>;

/**
 * @generated from message isuxportal.proto.resources.Notification.ClarificationMessage
 */
export declare type Notification_ClarificationMessage = Message<"isuxportal.proto.resources.Notification.ClarificationMessage"> & {
  /**
   * @generated from field: int64 clarification_id = 1;
   */
  clarificationId: bigint;

  /**
   * True when a clarification is sent from a team of notification recipient
   *
   * @generated from field: bool owned = 2;
   */
  owned: boolean;

  /**
   * True when a clarification was answered and have updated
   *
   * @generated from field: bool updated = 3;
   */
  updated: boolean;

  /**
   * True when a clarification was opened by admin
   *
   * @generated from field: bool admin = 4;
   */
  admin: boolean;
};

/**
 * Describes the message isuxportal.proto.resources.Notification.ClarificationMessage.
 * Use `create(Notification_ClarificationMessageSchema)` to create a new message.
 */
export declare const Notification_ClarificationMessageSchema: GenMessage<Notification_ClarificationMessage>;

/**
 * @generated from message isuxportal.proto.resources.Notification.TestMessage
 */
export declare type Notification_TestMessage = Message<"isuxportal.proto.resources.Notification.TestMessage"> & {
  /**
   * @generated from field: int64 something = 1;
   */
  something: bigint;
};